requirements.txt
django==4.2.27
python-dotenv==1.2.1